[fix] Change types of `Largesize` and `TileSize`
Because there was an error not to be able to convert the data.

+++ internal/dock/dock.go
@@ -19,7 +19,7 @@ type Plist struct {
 	PersistentApps              []PAItem `plist:"persistent-apps"`
 	PersistentOthers            []POItem `plist:"persistent-others"`
 	AutoHide                    bool     `plist:"autohide"`
-	Largesize                   float64  `plist:"largesize"`
+	Largesize                   int      `plist:"largesize"`
 	Loc                         string   `plist:"loc"`
 	Magnification               bool     `plist:"magnification"`
 	MinimizeToApplication       bool     `plist:"minimize-to-application"`
@@ -32,7 +32,7 @@ type Plist struct {
 	Region                      string   `plist:"region"`
 	ShowRecents                 bool     `plist:"show-recents"`
 	ShowAppExposeGestureEnabled bool     `plist:"showAppExposeGestureEnabled"`
-	TileSize                    float64  `plist:"tilesize"`
+	TileSize                    int      `plist:"tilesize"`
 	TrashFull                   bool     `plist:"trash-full"`
 	Version                     int      `plist:"version"`
 	WvousBlCorner               int      `plist:"wvous-bl-corner,omitempty"`
